+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loader/async/base.rs
