+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loader/async/cooperative.rs
@@ -1,12 +1,15 @@
-use crate::matmul::components::global::single_stage::loader::{
-    AsyncLoadingStrategy, CopyMechanism,
+use crate::matmul::components::{
+    global::{
+        tensor_view::{TensorReader, Window},
+        GlobalConfig, LoadingValidation,
+    },
+    stage::StridedTilingLayout,
+    Ident, InvalidConfigError, MatrixLayout,
 };
-use crate::matmul::components::global::tensor_view::{TensorReader, Window};
-use crate::matmul::components::global::{GlobalConfig, LoadingValidation};
-use crate::matmul::components::stage::StridedTilingLayout;
-use crate::matmul::components::{Ident, InvalidConfigError, MatrixLayout};
 use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
+
+use super::{AsyncLoadingStrategy, CopyMechanism};
 
 #[derive(CubeType, Clone, Copy)]
 /// Loads global memory into the stage without modification,
