+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loader/async/mod.rs
@@ -1,0 +1,7 @@
+mod base;
+mod cooperative;
+mod cyclic;
+
+pub use base::*;
+pub use cooperative::*;
+pub use cyclic::*;
