+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loader/mod.rs
@@ -1,7 +1,2 @@
-mod r#async;
-mod strategy;
-mod sync;
-
-pub use r#async::*;
-pub use strategy::*;
-pub use sync::*;
+pub mod r#async;
+pub mod sync;
