--- crates/cubecl-linalg/src/matmul/components/global/single_stage/loader/strategy/mod.rs
@@ -1,7 +1,0 @@
-mod cyclic_loading;
-mod stage_window_loading;
-mod tilewise_loading;
-
-pub use cyclic_loading::*;
-pub use stage_window_loading::*;
-pub use tilewise_loading::*;
