+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loader/sync/base.rs
