+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loader/sync/mod.rs
@@ -1,0 +1,7 @@
+mod base;
+mod cyclic;
+mod tilewise;
+
+pub use base::*;
+pub use cyclic::*;
+pub use tilewise::*;
