+++ crates/cubecl-linalg/src/matmul/kernels/matmul/algorithm/simple_barrier.rs
@@ -4,7 +4,7 @@ use std::marker::PhantomData;
 
 use crate::matmul::components::{
     batch::{self, CubeCountDispatch, CubeDispatch},
-    global::{self, single_stage::loader::WindowCooperativeLoading},
+    global::{self, single_stage::loader::r#async::WindowCooperativeLoading},
     stage, tile, MatmulProblem, MatmulSelection,
 };
